% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartStyleDataPrior.R
\name{bartStyleDataPrior}
\alias{bartStyleDataPrior}
\title{Estimate global shape and rate hyperparameters in BART-like fashion}
\usage{
bartStyleDataPrior(
  boundary,
  data_points,
  N_LEARNERS,
  MESH_REF,
  log_linear = FALSE
)
}
\arguments{
\item{MESH_REF}{}
}
\value{
A numeric vector containing the shape and rate hyperparameters
}
\description{
Uses a similar logic to that in Lamprinakou et. al (2023) to do a data
driven prior. It first tessellates the space (instead of partitioning via grid),
then treats the MLE of the intensity of each polygon divided by the number of learners
(or to the (1/number of learners) power if log_linear is TRUE)
as a 'sample'. The shape is set to the mean of the samples squared divided by the sample variance,
and the rate is set to the mean of the samples divided by the sample variance.
}
