% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityBASTFit.R
\name{DensityBASTFit_lpp}
\alias{DensityBASTFit_lpp}
\title{Fit DensityBAST model on linear point patterns}
\usage{
DensityBASTFit_lpp(data_matrix, hyperpars = NULL)
}
\arguments{
\item{data_matrix}{An object representing the linear network point pattern (e.g., an lpp object).}

\item{hyperpars}{A list of hyperparameters for the model. If NULL, default hyperparameters are used:
\itemize{
  \item N_LEARNERS = 16
  \item MESH_REF = 300
  \item MCMC = 2000
  \item BURNIN = 1000
  \item THIN = 1
  \item shape_hp = 1/16
  \item rate_hp = 1/16
  \item lambda_k_hp = 20
  \item max_clusters = 40
  \item inv_empirical_var = 1/16
  \item seed = NULL
  \item data_prior = TRUE
  \item normalized = FALSE
  \item hot_init = FALSE
  \item random_scan = FALSE
  \item detailed_output = TRUE
  \item subsetting = "thinning"
  \item thinning_ratio = 1
  \item parallel_cores = 16
}}
}
\value{
A list containing the fitted DensityBAST model.
}
\description{
Fits the DensityBAST model for point patterns on linear networks using an unconditioned density approach.
}
