% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBASTunblocked.R
\name{densityBASTunblocked}
\alias{densityBASTunblocked}
\title{Perform BASTION Density Estimation}
\usage{
densityBASTunblocked(
  learners,
  MCMC,
  BURNIN,
  THIN,
  shape_hp = 0.5,
  rate_hp,
  lambda_k_hp,
  max_clusters,
  seed = NULL,
  data_prior = FALSE,
  normalized = FALSE,
  hot_init = FALSE,
  random_scan = FALSE,
  detailed_output = FALSE,
  graph_detailed_output = FALSE,
  parallel_cores = NULL
)
}
\arguments{
\item{MCMC}{An integer, the number of iterations for the MCMC}

\item{BURNIN}{An integer, the number of initial iterations for the MCMC before recording output}

\item{THIN}{An integer, specifies the thinning interval}

\item{shape_hp}{Shape hyperparameter for the gamma prior. By default it is 1/2}

\item{rate_hp}{Rate hyperparameter for the gamma prior. Ignored if data_prior = TRUE}

\item{lambda_k_hp}{Hyperparameter for poisson prior on the number of clusters}

\item{max_clusters}{An integer, the maximum number of clusters for each weak learner}

\item{seed}{Optional, value to feed to R's set.seed function}

\item{data}{An object of class sfc_POINT; a list of the coordinates of observed data locations}

\item{boundary}{An object of class sfg POLYGON; the boundary which defines where the data can lie}

\item{N_LEARNERS}{An integer, the number of weak learners to use when training the model}

\item{MESH_REF}{An integer, the number of reference nodes in each weak learner's mesh. More reference nodes leads to a finer mesh}
}
\value{
A list containing two elements:
\item{lambda_output}{The piecewise constants corresponding to each cluster in each learner}
\item{membership_output}{The cluster memberships of ecah data point in each learner}
}
\description{
Perform BASTION Density Estimation
}
