% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensitySamples.R
\name{getDensitySamples}
\alias{getDensitySamples}
\title{Get matrix of posterior samples for a given grid}
\usage{
getDensitySamples(
  densityBASTOutput,
  gridPOINTs,
  intensity = TRUE,
  log_linear = FALSE
)
}
\arguments{
\item{densityBASTOutput}{The return object from the densityBAST function}

\item{gridPOINTs}{An sf object containing each grid point in its own row in the geometry column}

\item{intensity}{Optional; defaults to TRUE. A boolean value specifying whether to return the intensity function values (TRUE) or normalize the values so that each posterior sample is a density (FALSE)}
}
\value{
A matrix with dimension (number of grid points) x (number of posterior samples); each row corresponds to a grid point and each column is a posterior sample for that grid point.
}
\description{
Get matrix of posterior samples for a given grid
}
