% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinearDensitySamples.R
\name{getLinearDensitySamples}
\alias{getLinearDensitySamples}
\title{Get matrix of posterior samples for a given grid}
\usage{
getLinearDensitySamples(
  densityBASTOutput,
  eval_lpp,
  intensity = TRUE,
  log_linear = FALSE
)
}
\arguments{
\item{densityBASTOutput}{The return object from the densityBAST function}

\item{eval_lpp}{A spatstat.linnet::lpp object whose linear network is equivalent to `as.linnet(densityBASTOutput$data)`}

\item{intensity}{Optional; defaults to TRUE. A boolean value specifying whether to return the intensity function values (TRUE) or normalize the values so that each posterior sample is a density (FALSE)}
}
\value{
A matrix with dimension (number of evaluation points) x (number of posterior samples); each row corresponds to a grid point and each column is a posterior sample for that point.
}
\description{
Get matrix of posterior samples for a given grid
}
