% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotLearnersMeans.R
\name{ggplotLearnersMeans}
\alias{ggplotLearnersMeans}
\title{Generate a list of ggplot's for each weak learner}
\usage{
ggplotLearnersMeans(
  densityOutput,
  full_data,
  partitions = NULL,
  plotting_limit = 12
)
}
\arguments{
\item{densityOutput}{The output of the densityBAST function}

\item{data}{Optional; if a single}
}
\value{
A list of length N_LEARNERS where each element is a ggplot of the posterior mean (considering only the samples in the corresponding learner)
}
\description{
Generate a list of ggplot's for each weak learner
}
