% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDelauneyMeshes.R
\name{makeDelauneyMeshes}
\alias{makeDelauneyMeshes}
\title{Generate list of random Delauney Triangulation meshes on a boundary}
\usage{
makeDelauneyMeshes(
  boundary,
  N_LEARNERS,
  MESH_REF,
  MAX_ATTEMPTS = 100,
  eps = 1e-07
)
}
\arguments{
\item{eps}{}
}
\description{
Generate list of random Delauney Triangulation meshes on a boundary
}
