% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLearners.R
\name{makeLearners}
\alias{makeLearners}
\title{Generate learners for fitting densityBAST}
\usage{
makeLearners(meshes, data)
}
\arguments{
\item{meshes}{List of meshes, as returned by makeVoronoiMeshes or makeDelauneyMeshes}

\item{data}{An object of class sfc_POINT; a list of the coordinates of observed data locations}
}
\description{
Generate learners for fitting densityBAST
}
