% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLinearLearners.R
\name{makeLinearLearners}
\alias{makeLinearLearners}
\title{Title}
\usage{
makeLinearLearners(linenetwork_lpp, N_LEARNERS, MESH_REF, usepoisson = TRUE)
}
\arguments{
\item{linenetwork_lpp}{A spatstat.linnet::lpp class containing the observed points and line geometry}

\item{N_LEARNERS}{Number of weak learners}

\item{MESH_REF}{Number of mesh reference nodes per weak learner}

\item{usepoisson}{Boolean, whether to simulate reference nodes from a poisson point process instead of uniformly}
}
\description{
Title
}
