% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVoronoiMeshes.R
\name{makeVoronoiMeshes}
\alias{makeVoronoiMeshes}
\title{Generate list of random Voronoi meshes on a boundary}
\usage{
makeVoronoiMeshes(boundary, N_LEARNERS, MESH_REF, max_area_ratio = NULL)
}
\arguments{
\item{MESH_REF}{}
}
\value{
A list of length N_LEARNERS where each element is a collection of sf polygons
}
\description{
Generate list of random Voronoi meshes on a boundary
}
