% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeDensityFit.R
\name{summarizeDensityFit}
\alias{summarizeDensityFit}
\title{Evaluate performance of densityBAST against ground truth}
\usage{
summarizeDensityFit(
  density_output,
  grid_df,
  log_linear = FALSE,
  printout = TRUE
)
}
\arguments{
\item{density_output}{The output of a densityBAST function}

\item{grid_df}{A 3 column data frame containing a grid of 'X', 'Y', and 'true_intens' true intensity values at those coordinates}

\item{log_linear}{Passed into getDensitySamples; whether to interpret the output as coming from a log-linear model}

\item{printout}{Whether or not to print results to the console}
}
\value{
A list of containing plot_list, error_list, and est_intense_list
}
\description{
This function prints out Root Integrated Square Error, Average Absolute Error for
both the posterior mean and median as point estimates for the estimated intensity
function; as well as with just the posterior median for the estimated density function
}
